requirements.txt
Django
djangorestframework
celery
celery[redis]
django-celery-beat
djangorestframework-simplejwt
Faker
redis
django_celery_results
django-celery-beat      
djangorestframework-jwt-refresh-token
rest-framework-simplejwt        
drf-registration                
flake8
isort